restructure things a hair so they're less all over the place
