+++ src/ts/main.ts
@@ -10,8 +10,8 @@ import 'simple-keyboard/build/css/index.css';
 import VoteStatus from './protocol/VoteStatus.js';
 import * as bootstrap from 'bootstrap';
 import MuteState from './protocol/MuteState.js';
-import { I18nStringKey, TheI18n } from './i18n.js';
-import { Format } from './format.js';
+import { I18nStringKey, TheI18n } from './i18n';
+import { Format } from './util';
 import AuthManager from './AuthManager.js';
 import dayjs from 'dayjs';
 import * as dompurify from 'dompurify';
