+++ src/ts/tests/format.test.ts
@@ -1,18 +1,18 @@
-import { Format } from '../format';
+import { Format } from '../util';
 
-test('a string without any format specifiers in it is unaltered', () => {
+test('a string without any format specifiers in it is left as is', () => {
 	expect(Format('Hello World')).toBe('Hello World');
 });
 
-test('formatting a string works', () => {
+test('formatting a string works correctly', () => {
 	expect(Format('Hello, {0}!', 'World')).toBe('Hello, World!');
 });
 
-test('a cut off format specifier throws', () => {
+test('a cut off format specifier is not accepted', () => {
 	expect(() => Format('a{0', 1)).toThrow('Cutoff/invalid format specifier');
 });
 
-test('a malformed format specifier throws', () => {
+test('a malformed format specifier is not accepted', () => {
 	expect(() => Format('a{-0}', 1)).toThrow('Malformed format specifier');
 	expect(() => Format('a{0-}', 1)).toThrow('Malformed format specifier');
 	expect(() => Format('a{0ab}', 1)).toThrow('Malformed format specifier');
@@ -24,6 +24,6 @@ test('a malformed format specifier throws', () => {
 	expect(() => Format('a{ 0 }', 1)).toThrow('Whitespace inside format specifier');
 });
 
-test("a OOB format specifier doesn't work", () => {
+test("an out-of-bounds format specifier doesn't work", () => {
 	expect(() => Format('a {37}', 1)).toThrow('Argument index out of bounds');
 });
