+++ src/ts/util/StringLike.ts
