+++ src/ts/util/format.ts
