+++ src/ts/util/index.ts
@@ -1,0 +1,4 @@
+import { Format } from "./format";
+import { StringLike } from "./StringLike";
+
+export { Format, StringLike };
